safari-600.7-branch/LayoutTests/platform/mac/accessibility/press-action-not-exposed-when-body-is-click-handler.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html id="html">
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="group1" role="group">
text1
<br>
<div role="text" id="text2">text2</div>

</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that if the body or html tags have click handlers, then non-control elements do not automatically have the press action on them.");

    if (window.accessibilityController) {

          function listener() { }

          var group = accessibilityController.accessibleElementById("group1");
          var text1 = group.childAtIndex(0).childAtIndex(0);
          var text2 = accessibilityController.accessibleElementById("text2");

          shouldBe("text1.role", "'AXRole: AXStaticText'");
          shouldBe("text2.role", "'AXRole: AXStaticText'");

          document.getElementById("html").addEventListener('click', listener, false);

          debug("When a click handler is on the HTML tag, AXPress should not be supported on static text children elements automatically.");
          shouldBeFalse("text1.isPressActionSupported()");
          shouldBeFalse("text2.isPressActionSupported()");

          document.getElementById("html").removeEventListener('click', listener, false);
          document.getElementById("body").addEventListener('click', listener, false);

          debug("When a click handler is on the BODY tag, AXPress should not be supported on static text children elements automatically.");
          shouldBeFalse("text1.isPressActionSupported()");
          shouldBeFalse("text2.isPressActionSupported()");

          document.getElementById("body").removeEventListener('click', listener, false);
          document.getElementById("group1").addEventListener('click', listener, false);

          debug("When a click handler is on a parent tag, AXPress should be supported on static text children elements automatically.");
          shouldBeTrue("text1.isPressActionSupported()");
          shouldBeTrue("text2.isPressActionSupported()");

          document.getElementById("group1").style.visibility = "hidden";
    }

</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
